requirements.txt
git+https://github.com/ophexor/yt-dl@master
git+https://github.com/pyrogram/pyrogram@master
TgCrypto
py-tgcalls
python-dotenv
youtube_search_python
requests
aiohttp
aiofiles
asyncio
youtube_search
search_engine_parser
ffmpeg-python
ffmpeg
Pillow
ujson
